INT-621 | Verify notification bundle referential integrity
When a Search operation is performed on a FHIR resource and that resource is not found, no error is returned. If a resource we expect to be in the notification bundle is not found, it will throw an error
