+++ orchestrator/careplancontributor/ehr/notifier_test.go
@@ -2,7 +2,6 @@ package ehr
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"github.com/SanteonNL/orca/orchestrator/lib/test"
 	"github.com/google/uuid"
@@ -83,7 +82,7 @@ func TestNotifyTaskAccepted(t *testing.T) {
 		{
 			name:          "error fetching task",
 			task:          primaryTask,
-			expectedError: errors.New("fetch error"),
+			expectedError: errors.New("failed to create task notification bundle: fetch error"),
 			setupMocks: func(client *test.StubFHIRClient, _ *MockServiceBusClient) {
 				client.Error = errors.New("fetch error")
 			},
@@ -124,121 +123,4 @@ func TestNotifyTaskAccepted(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestIsOfType(t *testing.T) {
-	type1 := "Questionnaire"
-	type2 := "QuestionnaireResponse"
-	type3 := "https://example.com/Questionnaire/123"
-	type4 := "https://example.com/QuestionnaireResponse/123"
-	type6 := "https://example.com/Questionnaire"
-	type5 := "Questionnaire/123"
-	tests := []struct {
-		name           string
-		valueReference *fhir.Reference
-		typeName       string
-		expected       bool
-	}{
-		{
-			name: "type matches directly",
-			valueReference: &fhir.Reference{
-				Type: &type1,
-			},
-			typeName: "Questionnaire",
-			expected: true,
-		},
-		{
-			name: "type does not match directly",
-			valueReference: &fhir.Reference{
-				Type: &type2,
-			},
-			typeName: "Questionnaire",
-			expected: false,
-		},
-		{
-			name: "reference matches with https prefix",
-			valueReference: &fhir.Reference{
-				Reference: &type3,
-			},
-			typeName: "Questionnaire",
-			expected: true,
-		},
-		{
-			name: "reference does not match with https prefix",
-			valueReference: &fhir.Reference{
-				Reference: &type4,
-			},
-			typeName: "Questionnaire",
-			expected: false,
-		},
-		{
-			name: "reference matches without https prefix",
-			valueReference: &fhir.Reference{
-				Reference: &type5,
-			},
-			typeName: "Questionnaire",
-			expected: true,
-		},
-		{
-			name: "reference does match without https prefix",
-			valueReference: &fhir.Reference{
-				Reference: &type5,
-			},
-			typeName: "Questionnaire",
-			expected: true,
-		},
-		{
-			name: "reference does match without value",
-			valueReference: &fhir.Reference{
-				Reference: &type6,
-			},
-			typeName: "Questionnaire",
-			expected: false,
-		},
-		{
-			name: "nil reference",
-			valueReference: &fhir.Reference{
-				Reference: nil,
-			},
-			typeName: "Questionnaire",
-			expected: false,
-		},
-		{
-			name: "trigger a compilation error",
-			valueReference: &fhir.Reference{
-				Reference: &type4,
-			},
-			typeName: "(",
-			expected: false,
-		},
-		{
-			name: "nil type and reference",
-			valueReference: &fhir.Reference{
-				Type:      nil,
-				Reference: nil,
-			},
-			typeName: "Questionnaire",
-			expected: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := isOfType(tt.valueReference, tt.typeName)
-			require.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func marshalToRawMessage(resource any) json.RawMessage {
-	marshal, err := json.Marshal(resource)
-	if err != nil {
-		panic(err)
-	}
-	message := json.RawMessage{}
-	err = json.Unmarshal(marshal, &message)
-	if err != nil {
-		panic(err)
-	}
-	return message
 }
